Update deps for http api
The main riverd module (riverd) had problems building due to
dependencies. Building riverd-http-api pulled in either which pulled in
monad-control 1.0.0.1, riverd needs happstack-server which wants
monad-control <= 0.4 so updating the deps on riverd-http-api to
explicitly pull in 0.3.3.0 allows all packages to be built (luckily
either has bounds on monad-control (>=0.3.2 && <1.1).

The build dependencies need to be updated across the project to prevent
this, perhaps a switch to using stackage would also be beneficial?